+++ pycricinfo/api/endpoints/raw.py
@@ -30,7 +30,7 @@ async def player(player_id: int = Path(description="The Player ID")):
 
 
 @router.get(
-    "/match/{match_id}/basic",
+    "/match/{match_id}",
     responses={status.HTTP_200_OK: {"description": "The basic match data"}},
     summary="Get basic match data from the '/events' API",
 )
@@ -43,9 +43,7 @@ async def match_basic(match_id: int = Path(description="The Match ID")):
     responses={status.HTTP_200_OK: {"description": "The basic match data"}},
     summary="Get a match's Team",
 )
-async def get_match_team(
-    match_id: int = Path(description="The Match ID"), team_id: int = Path(description="The Team ID")
-):
+async def match_team(match_id: int = Path(description="The Match ID"), team_id: int = Path(description="The Team ID")):
     response = requests.get(
         f"http://core.espnuk.org/v2/sports/cricket/leagues/0/events/{match_id}/competitions/{match_id}/competitors/{team_id}"
     ).json()
@@ -53,16 +51,111 @@ async def get_match_team(
 
 
 @router.get(
-    "/match/{match_id}/summary",
+    "/match/{match_id}/team/{team_id}/roster",
+    responses={status.HTTP_200_OK: {"description": "The basic match data"}},
+    summary="Get a match Team's roster",
+)
+async def match_team_roster(
+    match_id: int = Path(description="The Match ID"), team_id: int = Path(description="The Team ID")
+):
+    response = requests.get(
+        f"http://core.espnuk.org/v2/sports/cricket/leagues/0/events/{match_id}/competitions/{match_id}/competitors/{team_id}/roster"
+    ).json()
+    return response
+
+
+@router.get(
+    "/match/{match_id}/team/{team_id}/innings",
+    responses={status.HTTP_200_OK: {"description": "The basic match data"}},
+    summary="Get a match Team's innings",
+)
+async def match_team_all_innings(
+    match_id: int = Path(description="The Match ID"),
+    team_id: int = Path(description="The Team ID"),
+):
+    response = requests.get(
+        f"http://core.espnuk.org/v2/sports/cricket/leagues/0/events/{match_id}/competitions/{match_id}/competitors/{team_id}/linescores"
+    ).json()
+    return response
+
+
+@router.get(
+    "/match/{match_id}/team/{team_id}/innings/{innings}",
+    responses={status.HTTP_200_OK: {"description": "The basic match data"}},
+    summary="Get a match Team's innings",
+)
+async def match_team_innings(
+    match_id: int = Path(description="The Match ID"),
+    team_id: int = Path(description="The Team ID"),
+    innings: int = Path(description="The innings number"),
+):
+    response = requests.get(
+        f"http://core.espnuk.org/v2/sports/cricket/leagues/0/events/{match_id}/competitions/{match_id}/competitors/{team_id}/linescores/0/{innings}"
+    ).json()
+    return response
+
+
+@router.get(
+    "/series/{series_id}/match/{match_id}/team/{team_id}/statistics",
+    responses={status.HTTP_200_OK: {"description": "The basic match data"}},
+    summary="Get a match Team's statistics",
+)
+async def match_team_statistics(
+    series_id: int = Path(description="The Series ID"),
+    match_id: int = Path(description="The Match ID"),
+    team_id: int = Path(description="The Team ID"),
+):
+    response = requests.get(
+        f"http://core.espnuk.org/v2/sports/cricket/leagues/{series_id}/events/{match_id}/competitions/{match_id}/competitors/{team_id}/statistics"
+    ).json()
+    return response
+
+
+@router.get(
+    "/match/{match_id}/team/{team_id}/player/{player_id}/innings",
+    responses={status.HTTP_200_OK: {"description": "The basic match data"}},
+    summary="Get a match Team's innings",
+)
+async def match_player_all_innings(
+    match_id: int = Path(description="The Match ID"),
+    team_id: int = Path(description="The Team ID"),
+    player_id: int = Path(description="The Player ID"),
+):
+    response = requests.get(
+        f"http://core.espnuk.org/v2/sports/cricket/leagues/0/events/{match_id}/competitions/{match_id}/competitors/{team_id}/roster/{player_id}/linescores"
+    ).json()
+    return response
+
+
+@router.get(
+    "/series/{series_id}/match/{match_id}/team/{team_id}/player/{player_id}/innings/{innings}/statistics",
+    responses={status.HTTP_200_OK: {"description": "The basic match data"}},
+    summary="Get a match Team's innings",
+)
+async def match_player_innings(
+    series_id: int = Path(description="The Series ID"),
+    match_id: int = Path(description="The Match ID"),
+    team_id: int = Path(description="The Team ID"),
+    player_id: int = Path(description="The Player ID"),
+    innings: int = Path(description="The innings number"),
+):
+    response = requests.get(
+        f"http://core.espnuk.org/v2/sports/cricket/leagues/{series_id}/events/{match_id}/competitions/{match_id}/competitors/{team_id}/roster/{player_id}/linescores/0/{innings}/statistics/0"
+    ).json()
+    return response
+
+
+@router.get(
+    "/match_summary/{match_id}",
     responses={status.HTTP_200_OK: {"description": "The match summary"}},
     summary="Get a match summary",
 )
-async def scorecard(match_id: int = Path(description="The Match ID")):
+async def match_summary(match_id: int = Path(description="The Match ID")):
     return get_request(get_settings().routes.match_summary, params={"match_id": match_id}, base_route=BaseRoute.site)
 
 
 @router.get(
-    "/match/{match_id}/play_by_play",
+    "/match/{match_id}/play_by_play_page",
     responses={status.HTTP_200_OK: {"description": "The match summary"}},
     summary="Get a page of ball-by-ball data",
 )